Return of the Language Wrapper
When I ditched the LanguageWrapper component I broke some important functionality. @ifsimicoded, thanks for catching that and being persistent in explaining to me, the dunce. 💜#128 (comment)

This PR should put things back to how the were mostly. With a "clean" App.js and a LanguageWrapper that has all the i18n language switch and cookie setting logic.

I did leave Google Analytics in App.js since it feels like a separate concern outside of languages. That didn't seem to effect functionality, but please call me out when I'm wrong. 😄
